Struct glium_sdl2::SDL2Facade

pub struct SDL2Facade { /* fields omitted */ }

Facade implementation for an SDL2 window.

Methods

impl SDL2Facade

pub fn window(&self) -> &Window

pub fn window_mut(&mut self) -> &mut Window

pub fn draw(&self) -> Frame

Start drawing on the backbuffer.

This function returns a Frame, which can be used to draw on it. When the Frame is destroyed, the buffers are swapped.

Note that destroying a Frame is immediate, even if vsync is enabled.